fixed errors occuring due to events helpers being server actions
