+++ services/server/events.ts
@@ -184,26 +184,3 @@ const formatCategories = (edges) => {
     };
   }
 };
-
-type EventsResponse = Awaited<
-  ReturnType<typeof client.queries.eventsContentQuery>
->;
-
-type EventsData = EventsResponse["data"]["events"];
-
-export const getTestimonialCategories = (events: EventsData): string[] => {
-  return (
-    events.testimonialCategories?.map(
-      (category) => category.testimonialCategory.name
-    ) || []
-  );
-};
-
-export const getVideoCardProps = (events: EventsData): VideoCardType[] => {
-  return (
-    events.videos?.videoCards?.map<VideoCardType>((m) => ({
-      title: m.title,
-      link: m.link,
-    })) || []
-  );
-};
